Expand on the unit test for random data.
The random data now uses 100,000 entries and
each test scenario is now in its own test case.
In addition, the path to every child is checked
to ensure the tree is balanced on depth (however,
the nodes may be imbalanced due to insertion ordering
and the size of the nodes, which is normal in the case of
a bplus tree).